+++ AWSAppSyncClient/AuthInterceptor/APIKeyAuthInterceptor.swift
