+++ AWSAppSyncClient/AuthInterceptor/AppSyncJSONHelper.swift
