+++ AWSAppSyncClient/AuthInterceptor/CognitoUserPoolsAuthInterceptor.swift
