+++ AWSAppSyncClient/AuthInterceptor/IAMAuthInterceptor.swift
