+++ AWSAppSyncClient/Subscription/Connection/ConnectionPool/SubscriptionConnectionFactory.swift
@@ -14,15 +14,6 @@ protocol SubscriptionConnectionFactory {
     func connection(connectionType: SubscriptionConnectionType) -> SubscriptionConnection?
 }
 
-/// Protocol for the different connection pool
-protocol SubscriptionConnectionPool {
-
-    /// Get Connection based on the url and connection type
-    /// - Parameter url: url to connect to
-    /// - Parameter connectionType:
-    func connection(for url: URL, connectionType: SubscriptionConnectionType) -> SubscriptionConnection
-}
-
 enum SubscriptionConnectionType {
 
     case appSyncRealtime
